x509: split TorServiceDescHash code into separate files.
Also switch to `CertificateFingerprint` type vs `[]byte` for the
`TorServiceDescriptorHash.Hash` field so that it will be serialized as
hex and not base64.
